+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3975,7 +3975,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KWHAAG7VAMGEYSGCMPQ3YF6W"),
+  id: z.string().default("01KWHDQ8EXPM2GXAJWSXZCZGC6"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),