+++ packages/orq-rc/RELEASES.md
@@ -10489,3 +10489,13 @@ Based on:
 - [typescript v4.11.0-rc.71] packages/orq-rc
 ### Releases
 - [NPM v4.11.0-rc.71] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.71 - packages/orq-rc
+
+## 2026-07-02 12:46:30
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.787.0 (2.914.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.11.0-rc.72] packages/orq-rc
+### Releases
+- [NPM v4.11.0-rc.72] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.72 - packages/orq-rc